void* Core_GetIvalAddrByType(const core_types_t type); msg_t Core_Module_Update(const core_types_t type, systime_t timeout_milliseconds); RGBW rewrite for Core_Module_Update

+++ modules/RGBW/RGBW.h
@@ -8,7 +8,7 @@ typedef struct
 {
 	volatile uint16_t 	Red, Green, Blue, White;
 	volatile uint16_t 	Red_Set, Green_Set, Blue_Set, White_Set;
-	volatile uint8_t	Rise_Time_Sec;
+	volatile uint16_t	Rise_Time_Sec;
 	volatile uint8_t	Max_Delay_Sec;
 	volatile systime_t 	Correction_24H_Sec;
 } RGBW_Inner_Val;
